food_ordering/requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0